set default value for EDXAPP_PLATFORM_NAME
The variable is set in `playbooks/roles/edxapp/defaults/main.yml` but
in the test suite, with the version of ansible used there, it appears
that it is not reliably available when the nginx role runs, resulting
in an exception and failed test. This doesn't seem to be an issue in
production, but is breaking the tests. This just sets a default to
help the test pass.